Port X page methods: page_sidebar, page_fillable, and page_navbar enhancements
